requirements.txt
torch
torchvision
timm
transformers
opencv-python
numpy
scipy
pillow
open3d
networkx
flask
requests
tenacity
tiktoken
openai
termcolor
python-dotenv
matplotlib
pandas
protobuf
PyYAML
tqdm
rich